tests/load/ refactor
